+++ src/main/java/frc/robot/CommandComposer.java
@@ -2,7 +2,7 @@ package frc.robot;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
-import frc.robot.commands.DriveDistanceCommand;
+import frc.robot.commands.Drive.DriveDistanceCommand;
 import frc.robot.subsystems.DriveSubsystem;
 
 public class CommandComposer {
