+++ src/main/java/frc/robot/commands/Drive/SetSteeringCommand.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.Drive;
 
 import edu.wpi.first.math.MathUtil;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
@@ -10,12 +10,12 @@ import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import frc.robot.subsystems.DriveSubsystem;
 
-public class SetSteering extends Command {
+public class SetSteeringCommand extends Command {
 	private DriveSubsystem m_subsystem;
 	private double m_angle;
 
 	/** Creates a new SetSteering. */
-	public SetSteering(DriveSubsystem subsystem, double angle) {
+	public SetSteeringCommand(DriveSubsystem subsystem, double angle) {
 		m_subsystem = subsystem;
 		m_angle = angle;
 		addRequirements(subsystem);
@@ -23,9 +23,9 @@ public class SetSteering extends Command {
 
 	public static Command getCalibrationCommand(DriveSubsystem subsystem) {
 		return new SequentialCommandGroup(
-				new SetSteering(subsystem, 45),
-				new SetSteering(subsystem, 0),
-				new SetSteering(subsystem, 45));
+				new SetSteeringCommand(subsystem, 45),
+				new SetSteeringCommand(subsystem, 0),
+				new SetSteeringCommand(subsystem, 45));
 	}
 
 	// Called when the command is initially scheduled.
